requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
Django==3.2.15
django-filter==2.4.0
djangorestframework==3.12.2
executing==1.2.0
ipdb==0.13.11
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.7.1
wcwidth==0.2.5
